refactor: Replace codes in loader into index instead
