+++ package.json
@@ -22,7 +22,7 @@
     "lint:check": "eslint . --ext .ts",
     "build:check": "tsc -p tsconfig.json",
     "build": "rimraf dist && node build.js",
-    "start": "node ./dist/loader.js"
+    "start": "node ./dist/index.js"
   },
   "devDependencies": {
     "@sentry/cli": "^2.31.2",
